+++ packages/presets/src/theme/recipes/menu.ts
@@ -36,12 +36,18 @@ export const menu = defineSlotRecipe({
       borderRadius: 'l2',
       borderWidth: '1px',
       boxShadow: 'lg',
-      _hidden: {
-        display: 'none',
-      },
       display: 'flex',
       flexDirection: 'column',
       outline: 'none',
+      _hidden: {
+        display: 'none',
+      },
+      _open: {
+        animation: 'fadeIn 0.25s ease-out',
+      },
+      _closed: {
+        animation: 'fadeOut 0.2s ease-out',
+      },
     },
     itemGroup: {
       display: 'flex',