+++ packages/presets/src/theme/recipes/number-input.ts
@@ -1,18 +1,49 @@
 import { numberInputAnatomy } from '@ark-ui/react'
 import { defineSlotRecipe } from '@pandacss/dev'
 
+const trigger = {
+  alignItems: 'center',
+  borderColor: 'border.emphasized',
+  cursor: 'pointer',
+  display: 'inline-flex',
+  justifyContent: 'center',
+  transitionDuration: 'normal',
+  transitionProperty: 'background, border-color, color, box-shadow',
+  transitionTimingFunction: 'default',
+  color: 'fg.emphasized',
+  _hover: {
+    color: 'fg.default',
+    background: 'bg.subtle',
+  },
+  _disabled: {
+    color: 'fg.disabled',
+    cursor: 'not-allowed',
+    _hover: {
+      background: 'transparent',
+      color: 'fg.disabled',
+    },
+  },
+}
+
 export const numberInput = defineSlotRecipe({
   className: 'numberInput',
   description: 'A number input style',
   slots: numberInputAnatomy.keys(),
   base: {
     root: {
+      display: 'flex',
+      flexDirection: 'column',
+      gap: '1.5',
+    },
+    control: {
       borderColor: 'border.emphasized',
       borderRadius: 'l2',
       borderWidth: '1px',
-      display: 'flex',
+      display: 'grid',
+      divideX: '1px',
+      gridTemplateColumns: '1fr 32px',
+      gridTemplateRows: '1fr 1fr',
       overflow: 'hidden',
-      width: 'full',
       transitionDuration: 'normal',
       transitionProperty: 'border-color, box-shadow',
       transitionTimingFunction: 'default',
@@ -21,23 +52,18 @@ export const numberInput = defineSlotRecipe({
         boxShadow: 'accent',
       },
     },
-    control: {
-      borderLeftWidth: '1px',
-      borderColor: 'border.emphasized',
-      display: 'flex',
-      flexDirection: 'column',
-    },
     input: {
       outline: 'none',
       background: 'transparent',
       width: 'full',
+      gridRow: '2',
     },
-    decrementTrigger: {
-      borderRadius: '0',
+    label: {
+      color: 'fg.emphasized',
+      fontWeight: 'medium',
     },
-    incrementTrigger: {
-      borderRadius: '0',
-    },
+    decrementTrigger: { ...trigger, borderTopWidth: '1px' },
+    incrementTrigger: trigger,
   },
   defaultVariants: {
     size: 'md',
@@ -45,27 +71,36 @@ export const numberInput = defineSlotRecipe({
   variants: {
     size: {
       md: {
-        root: {
+        control: {
           ps: '3',
           h: '10',
           minW: '10',
           fontSize: 'md',
         },
+        label: {
+          textStyle: 'sm',
+        },
       },
       lg: {
-        root: {
+        control: {
           ps: '3.5',
           h: '11',
           minW: '11',
           fontSize: 'md',
         },
+        label: {
+          textStyle: 'sm',
+        },
       },
       xl: {
-        root: {
+        control: {
           ps: '4',
           h: '12',
           minW: '12',
           fontSize: 'lg',
+        },
+        label: {
+          textStyle: 'md',
         },
       },
     },