+++ website/src/components/marketing/examples/e-commerce/quantity-picker.tsx
@@ -1,35 +1,26 @@
 import { FiChevronDown, FiChevronUp } from 'react-icons/fi'
-import { Divider, Stack } from 'styled-system/jsx'
-import { IconButton } from '~/components/ui/icon-button'
-import { Label } from '~/components/ui/label'
 import {
   NumberInput,
   NumberInputControl,
   NumberInputDecrementTrigger,
   NumberInputIncrementTrigger,
   NumberInputInput,
+  NumberInputLabel,
 } from '~/components/ui/number-input'
 
 export const QuantityPicker = () => {
   return (
-    <Stack gap="1.5">
-      <Label htmlFor="quantity">Quantity</Label>
-      <NumberInput min={1} max={5} defaultValue="1" size="xl">
-        <NumberInputInput id="quantity" />
-        <NumberInputControl>
-          <NumberInputIncrementTrigger asChild>
-            <IconButton size="xs" variant="tertiary" aria-label="Increment">
-              <FiChevronUp />
-            </IconButton>
-          </NumberInputIncrementTrigger>
-          <Divider />
-          <NumberInputDecrementTrigger asChild>
-            <IconButton size="xs" variant="tertiary" aria-label="Decrement">
-              <FiChevronDown />
-            </IconButton>
-          </NumberInputDecrementTrigger>
-        </NumberInputControl>
-      </NumberInput>
-    </Stack>
+    <NumberInput min={1} max={5} defaultValue="1" size="xl">
+      <NumberInputLabel>Quantity</NumberInputLabel>
+      <NumberInputControl>
+        <NumberInputInput />
+        <NumberInputIncrementTrigger>
+          <FiChevronUp />
+        </NumberInputIncrementTrigger>
+        <NumberInputDecrementTrigger>
+          <FiChevronDown />
+        </NumberInputDecrementTrigger>
+      </NumberInputControl>
+    </NumberInput>
   )
 }